Core/Globals: use proper type for smallint
(cherry picked from commit c9b730c767072225fdc2142908455c1d26e91fa8)

+++ src/server/game/Globals/ObjectMgr.cpp
@@ -7501,7 +7501,7 @@ void ObjectMgr::LoadGameObjectOverrides()
         }
 
         GameObjectOverride& gameObjectOverride = _gameObjectOverrideStore[spawnId];
-        gameObjectOverride.Faction = fields[1].GetUInt32();
+        gameObjectOverride.Faction = fields[1].GetUInt16();
         gameObjectOverride.Flags = fields[2].GetUInt32();
 
         if (gameObjectOverride.Faction && !sFactionTemplateStore.LookupEntry(gameObjectOverride.Faction))
